+++ test/mocha/custom.js
@@ -1,6 +1,7 @@
 var chai = require('chai');
 var expect = chai.expect;
 var critr = require('../../src/critr');
+var data = null;
 
 describe('Custom Ops:', function () {
     describe('register op', function () {
@@ -36,6 +37,36 @@ describe('Custom Ops:', function () {
             })).to.be.false;
 
             expect(critr.test({ x: 1 }, { x: { $test: 0 }})).to.be.true;
+        });
+    });
+
+    describe('register stage op', function () {
+        beforeEach(function () {
+            critr = new critr.Critr();
+            data = [{
+                name: 'john'
+            }];
+        });
+
+        it('should allow custom stage op registration', function () {
+            expect(critr.stage('$test', function (context) {
+                context.output(context.data[0]);
+            })).to.be.true;
+        });
+
+        it('should allow use of custom stage op', function (done) {
+            critr.stage('$test', function (context, next) {
+                context.output(context.param);
+                next();
+            });
+
+            critr.aggregate(data, [{
+                $test: { name: 'bob' }
+            }], function (result) {
+                expect(result).to.have.length(1);
+                expect(result[0]).to.have.property('name', 'bob');
+                done();
+            });
         });
     });
 });