+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `evaluate` method to public interface.
 - Code Climate badge to README.
 - get operation with `operator` method.
+- Custom stage operators using the `stage` method, similar to `operator` but for pipeline stage operators.
 
 ### Changed
 - The public interface into a type accessible via `<instance>.Critr`